Refactor into separate middleware
